Add zoom functionality for Windows
Without this menu item, increasing and decreasing the font size of the app does not work on Windows. It does work on Mac, with the regular Chrome shortcut keys (Cmd-+, Cmd--, and Cmd-0 to restore original size). When this menu item is added, font size can be changed on Windows with Ctrl-Shift-+ and Ctrl--.